chore: move scroll restoration and focus reset out of client.js
